move workspace crate to crate & update names
