+++ crates/semaphore/src/hash.rs
