+++ crates/semaphore/src/packed_proof.rs
