+++ crates/semaphore/src/protocol/authentication.rs
