+++ crates/semaphore/src/protocol/mod.rs
@@ -10,13 +10,13 @@ use ark_std::UniformRand;
 use color_eyre::Result;
 use ethers_core::types::U256;
 use once_cell::sync::Lazy;
-use poseidon::Poseidon;
 use rand::{thread_rng, Rng};
-use semaphore_depth_config::{get_depth_index, get_supported_depth_count};
-use semaphore_depth_macros::array_for_depths;
+use semaphore_rs_depth_config::{get_depth_index, get_supported_depth_count};
+use semaphore_rs_depth_macros::array_for_depths;
+use semaphore_rs_poseidon::Poseidon;
+use semaphore_rs_trees::{Branch, InclusionProof};
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
-use trees::Branch;
 use witness::Graph;
 
 use crate::circuit::zkey;
@@ -73,7 +73,7 @@ impl From<Proof> for ArkProof<Bn<Config>> {
 
 /// Helper to merkle proof into a bigint vector
 /// TODO: we should create a From trait for this
-fn merkle_proof_to_vec(proof: &trees::Proof<Poseidon>) -> Vec<Field> {
+fn merkle_proof_to_vec(proof: &InclusionProof<Poseidon>) -> Vec<Field> {
     proof
         .0
         .iter()
@@ -86,7 +86,7 @@ fn merkle_proof_to_vec(proof: &trees::Proof<Poseidon>) -> Vec<Field> {
 /// Generates the nullifier hash
 #[must_use]
 pub fn generate_nullifier_hash(identity: &Identity, external_nullifier: Field) -> Field {
-    poseidon::poseidon::hash2(external_nullifier, identity.nullifier)
+    semaphore_rs_poseidon::poseidon::hash2(external_nullifier, identity.nullifier)
 }
 
 #[derive(Error, Debug)]
@@ -108,7 +108,7 @@ pub enum ProofError {
 /// Returns a [`ProofError`] if proving fails.
 pub fn generate_proof(
     identity: &Identity,
-    merkle_proof: &trees::Proof<Poseidon>,
+    merkle_proof: &InclusionProof<Poseidon>,
     external_nullifier_hash: Field,
     signal_hash: Field,
 ) -> Result<Proof, ProofError> {
@@ -128,7 +128,7 @@ pub fn generate_proof(
 /// Returns a [`ProofError`] if proving fails.
 pub fn generate_proof_rng(
     identity: &Identity,
-    merkle_proof: &trees::Proof<Poseidon>,
+    merkle_proof: &InclusionProof<Poseidon>,
     external_nullifier_hash: Field,
     signal_hash: Field,
     rng: &mut impl Rng,
@@ -145,7 +145,7 @@ pub fn generate_proof_rng(
 
 fn generate_proof_rs(
     identity: &Identity,
-    merkle_proof: &trees::Proof<Poseidon>,
+    merkle_proof: &InclusionProof<Poseidon>,
     external_nullifier_hash: Field,
     signal_hash: Field,
     r: ark_bn254::Fr,
@@ -172,7 +172,7 @@ fn generate_proof_rs(
 
 pub fn generate_witness(
     identity: &Identity,
-    merkle_proof: &trees::Proof<Poseidon>,
+    merkle_proof: &InclusionProof<Poseidon>,
     external_nullifier_hash: Field,
     signal_hash: Field,
 ) -> Vec<Fr> {
@@ -201,7 +201,7 @@ pub fn generate_witness(
 
 /// Compute path index
 #[must_use]
-pub fn path_index(proof: &trees::Proof<Poseidon>) -> Vec<Field> {
+pub fn path_index(proof: &InclusionProof<Poseidon>) -> Vec<Field> {
     proof
         .0
         .iter()
@@ -244,7 +244,7 @@ pub fn verify_proof(
 mod test {
     use rand::SeedableRng as _;
     use rand_chacha::ChaChaRng;
-    use semaphore_depth_macros::test_all_depths;
+    use semaphore_rs_depth_macros::test_all_depths;
     use serde_json::json;
 
     use super::*;
