+++ crates/semaphore/src/util.rs
